requirements.txt
streamlit
pandas
scikit-learn
python-dotenv
openai